Refactor Family Garden page and related components
