+++ src/app/pdf/page.tsx
@@ -8,6 +8,7 @@ import { useFamilyVisionStore } from "../family-vision/family-vision-store";
 import { usePhilanthropyStore } from "../philanthropy/philanthropy-store";
 import { useContactStore } from "../contacts/contact-store";
 import { useFamilyGardenStore } from "../family-garden/family-garden-store";
+import { PersonTable } from "../family-garden/person-table";
 
 export default function PDFPage() {
   const questions = useDecisionTreeStore((s) => s.questions);
@@ -24,10 +25,8 @@ export default function PDFPage() {
   const guidelines = usePhilanthropyStore((s) => s.guidelines);
   const impactStatement = usePhilanthropyStore((s) => s.impactStatement);
 
-  const { growthRate, people } = useFamilyGardenStore((state) => ({
-    growthRate: state.growthRate,
-    people: state.people,
-  }));
+  const growthRate = useFamilyGardenStore((s) => s.growthRate);
+  const people = useFamilyGardenStore((s) => s.people);
 
   const handleClick = () => {
     window.print();
@@ -134,21 +133,13 @@ export default function PDFPage() {
             }
           </div>
         </Page>
-        <Page>
-          <h1>Family Garden</h1>
-          <p>Growth Rate: {(growthRate * 100).toFixed(2)}%</p>
-          {people.length > 0 ? (
-            people.map((person) => (
-              <div key={person.id}>
-                <h2>{person.name || `Person ${person.id}`}</h2>
-                <p>Begin Age: {person.beginAge}</p>
-                <p>Begin Amount: {person.beginAmount}</p>
-              </div>
-            ))
-          ) : (
-            <p>No people added yet.</p>
-          )}
-        </Page>
+        {people.map((person) => (
+          <Page key={person.name}>
+            <h1>Family Garden</h1>
+            <h2 className="">{person.name}</h2>
+            <PersonTable person={person} />
+          </Page>
+        ))}
       </div>
     </main>
   );
